fix: Allow again branches in git URLs (#171)
Change #166 (commit ffe63bf) changed resolution of commit-ish argument
in git URLs. The change broke resolution for branches. After normal
clone only main/master is available due to checkout. Any other branch
needs to be tracked through its respective remote, in this case
`origin`.

In the new behavior commit-ish argument is first assumed to be a tag
name or main/master branch. If the resolution fails it is assumed to be
branch from origin.

Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>